requirements.txt
numpy
pandas
matplotlib
opencv-python
tensorflow>=2.0.0
Pillow
scikit-learn
flask
werkzeug